requirements.txt
ccxt
flask
flask-cors
requests
pandas
numpy
ta
python-dotenv
gunicorn
yfinance
python-binance
alpha_vantage
pybit
openai
